- AJAX - Cookie - Regex - Liste mit eingesetzten Technologien

+++ checkout/overview.php
@@ -3,6 +3,7 @@
     $root = realpath($_SERVER["DOCUMENT_ROOT"]);
     
     include("$root/cagelovers/src/Initializer.php") ;
+    $_SESSION['navipoint'] = '';
     include("$root/cagelovers/templates/header.php");
     include("$root/cagelovers/src/db/ArticleSubDB.inc");
     
@@ -46,36 +47,50 @@
                     
                     <?php
                         
-                        foreach ($basket->items as &$value) {
-                            
-                            if(isset($value->modification)) {
-                                $currentArticleSub = new ArticleSubDB();
-                                $currentArticleSub = $currentArticleSub->getSpecificArticleSub($value->modification);
-                                $currentArticleSub = $currentArticleSub->fetch_object();
+                        if(isset($_SESSION['basket']) && isset($basket->items)) {
+                            foreach ($basket->items as &$value) {
+
+                                if(isset($value->modification)) {
+                                    $currentArticleSub = new ArticleSubDB();
+                                    $currentArticleSub = $currentArticleSub->getSpecificArticleSub($value->modification);
+                                    $currentArticleSub = $currentArticleSub->fetch_object();
+                                }
+
+                                echo '<tr>';
+
+                                echo '<td>' . $value->name . '</td>';
+                                echo '<td>' . $value->amount . '</td>';
+                                echo '<td>' . $value->prize . '.-</td>';
+                                if(isset($value->modification)) echo '<td>' . $currentArticleSub->Value . '</td>';
+                                else echo '<td></td>';
+                                echo '<td>' . $value->amount * $value->prize . '.-</td>';
+                                echo '<td class="last"><a href="/cagelovers/checkout/overview.php?delete='.$value->id.'" class="icon-delete glyphicon glyphicon-remove-circle"></a></td>'; 
+
+                                echo '</tr>';
                             }
-                            
-                            echo '<tr>';
-                            
-                            echo '<td>' . $value->name . '</td>';
-                            echo '<td>' . $value->amount . '</td>';
-                            echo '<td>' . $value->prize . '.-</td>';
-                            if(isset($value->modification)) echo '<td>' . $currentArticleSub->Value . '</td>';
-                            else echo '<td></td>';
-                            echo '<td>' . $value->amount * $value->prize . '.-</td>';
-                            echo '<td class="last"><a href="/cagelovers/checkout/overview.php?delete='.$value->id.'" class="icon-delete glyphicon glyphicon-remove-circle"></a></td>'; 
-                            
-                            echo '</tr>';
                         }
                     
                     ?>
                   </table>
                 
                 <div class="additional-info clearfix">
-                    <div class="prize-total">Preis total: <strong><?php echo $basket->getPrizeTotal(); ?>.-</strong></div>
+                    <div class="prize-total">Preis total: <strong>
+                        <?php
+                            if(isset($_SESSION['basket'])) {
+                                echo $basket->getPrizeTotal();
+                            } else {
+                                echo '0';
+                            }
+                            
+                        ?>.-</strong></div>
                 </div>
                 
                 <div class="additional-info clearfix">
-                    <a href="/cagelovers/checkout/payment.php" class="btn btn-success">Weiter</a>
+                    <?php
+                        if(isset($_SESSION['basket']) && isset($basket->items) && (count($basket->items) > 0)) {
+                            echo '<a href="/cagelovers/checkout/payment.php" class="btn btn-success">Weiter</a>';
+                        }
+                    ?>
                 </div>
                 
             </div>
